+++ src/collective/behavior/seo/behaviors/seo_fields.py
@@ -8,32 +8,59 @@ from plone.supermodel import model
 from zope.component import adapter
 from zope.interface import implementer
 from zope.interface import provider
+from plone.autoform import directives
 
+from ..interfaces import ISEOFieldsMarker
 
 @provider(IFormFieldProvider)
-class ISeoFields(model.Schema):
+class ISEOFields(model.Schema):
     """
     """
+    model.fieldset(
+            'seofields',
+            label=_(u'SEO'),
+            fields=(
+                'seo_title',
+                'seo_description'
+            ),
+        )
 
-    project = schema.TextLine(
-        title=_(u'Project'),
-        description=_(u'Give in a project name'),
-        required=False,
-    )
+    seo_title=schema.TextLine(
+        title=u"SEO Title",
+        description=_(u"This field is used in the web page <head> section "
+                      u"title instead of the title on the main tab."),
+        required=False
+        )
 
+    seo_description=schema.Text(
+        title=u"SEO Description",
+        description=_(u"This field is used as meta description field in <head> section and browser "
+                      u"tab instead of the description on the main tab."),
+        required=False
+        )
 
-@implementer(ISeoFields)
+@implementer(ISEOFieldsMarker)
 @adapter(IDexterityContent)
-class SeoFields(object):
+class SEOFields(object):
     def __init__(self, context):
         self.context = context
 
     @property
-    def project(self):
-        if hasattr(self.context, 'project'):
-            return self.context.project
+    def seo_title(self):
+        if hasattr(self.context, 'seo_title'):
+            return self.context.seo_title
         return None
 
-    @project.setter
-    def project(self, value):
-        self.context.project = value
+    @seo_title.setter
+    def seo_title(self, value):
+        self.context.seo_title = value
+    
+    @property
+    def seo_description(self):
+        if hasattr(self.context, 'seo_description'):
+            return self.context.seo_description
+        return None
+
+    @seo_description.setter
+    def seo_description(self, value):
+        self.context.seo_description = value