+++ CHANGES.rst
@@ -5,5 +5,5 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
-- Initial release.
+- Basic version, only for Title and Description
   [zestsoftware]